Uploading all the files into github
I am aware, that there is a code issue, but I am currently to mentally drained to even look at the code. I will figure out all the kinks of it tomorrow. And I am quite excited to do so, but until then I ensured that the code at least runs even though the output is incorrect for the time being.